kernel/sched: Disable GFS (Gentle Fair Sleepers)
Sleeper Fairness is a concept used by CFS which treat sleeping/waiting tasks as if they were in a run queue.
This implies tasks which spend most of the time waiting for an user input and such will get a fair share of
CPU when they need it. Disabling Gentle Fair Sleepers could improve UI responsiveness.

Signed-off-by: Gabriel2392 <gabriel824m@gmail.com>
Change-Id: I7b300dd9d19756bded478a5e857346dbb6cab13c
Signed-off-by: Chatur27 <jasonbright2709@gmail.com>