kthread: Add API to create kthreads bound to low-power CPUs
Some unimportant kthreads do not need high-performance and can safely be
bound to low-power CPUs for energy efficiency.

This commit adds an API to create kthreads bound to low-power CPUs. This
affinity cannot be altered by the kernel or userspace.

This extends the performance-critical kthread support introduced in
commit 51ae24deb56007eea83895509bbc58732558114a ("kernel: Add API to mark IRQs and kthreads as
performance critical").

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Chatur27 <jasonbright2709@gmail.com>