drivers/leds/leds-s2mu005: Custom sysfs node for flashlight
This creates a custom sysfs node located at sys/devices/virtual/camera/flash/torch_brightness_lvl
for the supported LED. This sysfs node can be used to control flashlight brightness on
AOSP ROMs or recoveries.

By default, the brightness of flashlight is set to maximum, that is, it will fix low
flashlight issue on AOSP ROMs, GSIs and recoveries.

This sysfs node can also override the stock behaviour of ONEUI's flashlight controller
(only if uncommented in the driver code).

The flashlight brightness range is from 0 to 10.

Change-Id: I17eaa0a253492d211210c91fa4281d134e1c3328
Signed-off-by: Chatur27 <jasonbright2709@gmail.com>