drivers: soc: acpm: Only optimize the ACPM module for size Optimizing the cokernel pilation for performance introduces more issues with the ACPM driver. Very similar to its earlier issues with the kernel not being able to "access kernel address".
I'm seriously tired of playing games with Exynos ACPM and its poorly-coded source. As soon as we enable additional optimizations it just causes so much issues.

Since we're not experiencing issues when it's optimized for size, let's just make the whole module *permanently* optimized for size regardless. This will allow us to optimize the rest of the kernel for performance but leave this out.

We might also do this to CAL-IF if it also causes issues.

Signed-off-by: John Vincent <git@tensevntysevn.cf>