media: exynos: fimc-is2: use monotonic time to get the "boot" timestamp. Samsung's Exynos camera HALs for exynos7870 devices, and others, for Android use the boot timestamp obtained using boottime, rather than the timestamp obtained using monotonic time. In Android's libcameraservice, the clock offset between boottime and monotonic time is calculated only if the timestamp source is reported as realtime by the camera HAL for conversion of the timestamp from boottime to monotonic time. (if condition for that for context: https://github.com/LineageOS/android_frameworks_av/blob/41f3df68de3b2075027070a7555679ad004794fb/services/camera/libcameraservice/device3/Camera3Device.cpp#L319) The camera HALs do not report the source of the timestamp as realtime, and therefore, the offset is never calculated. This causes an issue where videos that are taken have a duration that is lengthened by the time that is spent suspended/in deep sleep, as boottime includes any time that is spent suspended.
To fix that issue, use monotonic time for the boot (or now "boot") timestamp
in the fimc-is2 driver, by replacing a call to fimc_is_get_timestamp_boot(),
which returns a timestamp calculated from boottime, to get the boot timestamp
with a call to fimc_is_get_timestamp(), which returns a timestamp calculated
from monotonic time.

This change makes the patch titled:
	"camera: Allow to use boottime as timestamp reference"
as was merged to LineageOS' android_frameworks_av repository on branch
lineage-18.1, and also setting the TARGET_CAMERA_BOOTTIME_TIMESTAMP
variable to true along with it in BoardConfig.mk in the device tree, no
longer necessary.

Signed-off-by: TALU <taluemail272@gmail.com>
Change-Id: Ifb4568de5fe41f467eb81b7a8a4f29a2e3f7b4a7
Signed-off-by: SamarV-121 <samarvispute121@pm.me>