cpuidle: Expose use_deepest_state as a sysfs attribute The use_deepest_state property introduced in commit 98ddb63d7efdb is not currently modified by anything. However, this may be a useful property to control from userspace which is aware of the current context. This also improves the ability to debug the status of use_deepest_state if and when we do start mutating this property from kernel-space.
This attribute is read-only when cpuidle_sysfs_switch is disabled and
mutable if it is enabled, similar to how current_governor[_ro] behaves.

The current implementation of this attribute assumes that all CPUs will
always have the same use_deepest_state value for simplicity. This may be
revisited in the future if we decide to deviate from this initial
assumption, e.g. enabling use_deepest_state for only one CPU cluster on
heterogeneous systems.

Test: writing to and reading from use_deepest_state works as expected
Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: Yasir-Siddiqui <www.mohammad.yasir.s@gmail.com>
Signed-off-by: sohamxda7 <sensoham135@gmail.com>
Signed-off-by: Prashant-1695 <prashant33968@gmail.com>
Signed-off-by: ImPrashantt <prashant33968@gmail.com>